Generalize hidden-selection-surface capture
In preparation for fullscreen capture
