+++ include/freezeframe.h
@@ -10,7 +10,6 @@ void freezeframe_capture_start_assume_callback_set(struct scran_output *st_outpu
 void freezeframe_capture_refresh(struct scran_output *st_output, freezeframe_callback callback);
 
 void freezeframe_hide_surface(struct scran_output *st_output);
-void freezeframe_unhide_selection_surface(struct scran_output *st_output);
 
 void freezeframe_surface_update_scale_size_viewport(struct scran_output *st_output);
